LayoutTests/canvas/philip/tests/2d.imageData.object.undefined-expected.txt

< [index] >
2d.imageData.object.undefined
ImageData.data converts undefined to 0
References: 2d.pixelarray.modify
Actual output:
Passed
